+++ src/contents/ui/ColorButton.qml
@@ -1,0 +1,20 @@
+/**
+ * Weekday Widget for KDE
+ *
+ * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
+ * @copyright 2020 Marcin Orlowski
+ * @license   http://www.opensource.org/licenses/mit-license.php MIT
+ * @link      https://github.com/MarcinOrlowski/weekday-plasmoid
+ */
+
+import QtQuick 2.0
+import QtQuick.Controls 2.3
+import QtQuick.Layouts 1.1
+import org.kde.kquickcontrols 2.0 as KQControls
+
+KQControls.ColorButton {
+//	id: customColorsTodayFg
+	showAlphaChannel: true
+	dialogTitle: ''
+	Layout.alignment: Qt.AlignHCenter
+}