+++ src/contents/ui/LabelCenter.qml
@@ -1,0 +1,18 @@
+/**
+ * Weekday Widget for KDE
+ *
+ * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
+ * @copyright 2020 Marcin Orlowski
+ * @license   http://www.opensource.org/licenses/mit-license.php MIT
+ * @link      https://github.com/MarcinOrlowski/weekday-plasmoid
+ */
+
+import QtQuick 2.0
+import QtQuick.Layouts 1.1
+import QtQuick.Controls 2.3
+import org.kde.plasma.components 3.0 as PlasmaComponents
+
+PlasmaComponents.Label {
+	text: ''
+	Layout.alignment: Qt.AlignHCenter
+}